xts5: Use pkg-config to get X libraries
We need to link the xts libraries and test applications with X libraries.
This was being defined in common.mk as *SYSLIBS, but now the convention
is to use the <var>_{CFLAGS,LIBS} used by PKG_CHECK_MODULES.

A couple of the existing CFLAGS variables needed to be renamed to avoid
conflicting with the pkg-config variables. This is all still pretty messy
in common.mk, and should be cleaned up eventually.