tcc: Don't require a journal or a results directory
When the user passes an empty string for the journal argument (-j) or
the results directory argument (-i), don't create them. This allows tcc
to be run in more of a single shot mode since the logging is still done
in the tet_xres file.

The fullpath() code has been modified to use the current directory when
the argument is NULL since it is sometimes called with the results
directory name.