tcc: Exit with a failure code if any of the tests failed
tcc was always exiting successfully so long as all the processes it was
governing were completing. However, we want it to fail if any of the
processes (tests) fail.

Changes have been made throughout the code to pass up results of tests
to tcc. The tet_addresult function allows multiple integer return codes
to be handled by prefering failing codes.