Turn the main loop code into a library, libtcmmain
Rather than compiling tcm.c and dtcm.c for every test program, just
build libtcmmain.a once and link it in. The same goes for tcmchild.c and
libtcmchild.a.

This also allows for an automake time saver since we can move tcm into
the LIBS definition and then the program is a single source file of the
same name. Automake defaults to using this source file without needing
to define _SOURCES.