xts5: Run all Xproto tests separately
Instead of giving tcc one big scenario file and waiting for it to
complete, run tcc with a single test/scenario through automake's test
framework. This way, we get a single PASS/FAIL/SKIP for each test.

Unfortunately, this means that the scenarios are duplicated for a third
time in the repo. Also, nearly all the Xproto tests fail for one reason
or another on my machine. Part of this might be too much strictness in
tcc's exit status.