mc: Use unique temporary files to avoid conflicts
mc was using non-unique temporary files to store generated content
before copying all the files together into the .c file. When running
multiple mc processes, they all try to use the same files, which results
in strange errors. Now the pid is suffixed to the temporary file to give
a reasonable sense that it will be unique.

The mc pattern rule has been changed back to operate on files instead of
stdin/stdout since that change was supposed to fix this issue.