Compile objects in the same subdirectory as the sources
For the tests, we want to just have top-level Makefiles and build the
sources in their own subdirectory. Typically, automake would create the
object files at the top-level, but this can cause conflicts when the
source files have the same name.

Use the subdir-objects automake option and enable the necessary
AM_PROG_CC_C_O macro.