tet: Create results file from test program name
When the results file is always named "tet_xres", we can't have multiple
results files in the same directory. Instead, derive the results file
from the test program name. It would be best if the scenario name was
used, but I don't think that's available at the time the file is
created.