xts5: Execute tests by test name instead of program name
The test program names are confusing, and printing their names during
testing is not helpful. For instance, when running the XBlackPixel test,
make showed blckpxl.

This changes the tests to be scenario names from the top level tet_scen
file. Since the test is not a file or program, they need to be declared
PHONY. By using the tet_scen file, we no longer need the individual
.scen files.