mc: Do all processing in the source directory
To ensure that included files are resolved correctly and all temporary
files are created in a reasonable place, move to the directory of the
source file to do all processing. To resolve the output filename, it is
opened prior to changing directories.

The directory changing code is not portable, but someone else will have
to step up to make it portable to their platform.