tcc: Print a summary of the test results
Make tcc track the number and status of tests run in the scenario and
print a summary of results after processing all tests. The output mimics
automake's test driver.

The summary code makes use of snprintf. Anyone can feel free to make it
more portable, but it seemed like a PITA to me.