tcc: Print test suite and test case names in reports
tcc provides the test suite name (xts5) and test case name (e.g.
/Xproto/pAllocColor) to the test program via the environment. The test
program uses this information when outputting the results to stdout.

TET_TSNAME is a new global variable containing the test suite name
provided by tcc. This required a change to the config initialization to
only check if the value was an absolute path when necessary.

TET_TCNAME is a per-test-exec variable providing the test case name as
listed in the scenario file.