Use a default TET_ROOT when not in the environment
Requiring an environment variable to be set in order to get any behavior
is an obnoxious trait. Instead, a sane default TET_ROOT is built into
the programs that use it. Users such as "make check" that need more
control can continue using the environment variable.

The scripts are now substituted from make so that we get full variable
resolution.