tet: Allow for executables outside of TET_ROOT by default
tet is really built around the notion that all the source, programs, data
and output will all live in a single tree. However, often times the test
programs are built and/or installed in a separate location. This can be
worked around with various environment variables or command line options,
but it would be better to build this behavior in by default.

This introduces TET_EXEC_ROOT, which specifies the base directory for tet
test programs. The tests will be found in the test suite subdirectory of
TET_EXEC_ROOT. E.g., $TET_EXEC_ROOT/xts. The default here is to have
TET_ROOT=$datadir and TET_EXEC_ROOT=$libexecdir.