xts5: Use shared library modules to decrease test program size
Linking the xts libraries statically into each test program results in a
massive size for the collection of programs. The libraries have been
changed to private libtool modules that default to shared, resulting in
a much smaller installed size.

One consequence of linking to shared objects is that the linker errors
on undefined extern variables. The obsolete and unused startup routines
in linkstart.c had one of these, so they have been removed from the
build.

Next is to create the tet library shared.