tet: Fix "srcFile defined but not used" warnings.
These global arrays are used by the error-reporting macros to define the source
file.  I guess it's because using __FILE__ everywhere generated tons of copies
of the file path in .data on ancient compilers.  Nowadays, they end up with only
one copy in .rodata.  To facilitate that, make the arguments to the error
reporting functions const.  Propagate constness to eliminate additional
warnings.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>