Xlib13 XGrabButton: don't overrun the number of buttons configured.
NS_button contains the static button mappings (for whatever reason). Don't
try to access it nbuttons() time because that will return the number of
buttons configured on the device (usually 12). Instead, use the size of the
array we're traversing through.

Aside from running OOB on an array being considered impolite, it also gives
us BadValue when trying to XTestFakeButtonEvent a Button 0, which the test
suite doesn't like.

Fixes UNRESOLVED on test case 10/39.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>