Xlib13: fix XSetModifierMapping test 1/8.
XSetModifierMapping reports FAIL due to a different ordering of the
modifiers in the return value. The server keeps the modifiers in a
different format than used by the core protocol. The modifier map
returned by XGetModifierMapping(3) always has the keycodes in
ascending order for each modifier.

A straight input_map[i] == output_map[i] comparison is not enough if
the input map is not in order. Instead, all values for each modifier
need to be compared to find the right modifier.

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
[aplattner@nvidia.com: clarification comment added]
Reviewed-by: Aaron Plattner <aplattner@nvidia.com>
Signed-off-by: Aaron Plattner <aplattner@nvidia.com>