Fix a typo in the README.
"It is not be necessary".

I did notice that some of the tests, such as XCreateGlyphCursor, fail with
UNRESOLVED because XT_FONTPATH is "../xts5/fonts" so the server can't find the
font files.  I still need to figure out why that's happening.  The test runs
fine if the suite is installed, or if I modify tetexec.cfg to contain the
absolute path to the fonts.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>