xts-config.in: ":n" is a valid display name
When determining whether the connection is local and you can connect to the
server via TCP, don't try to parse out ":n.m" and then reconstruct it as
"hostname:n.m", because ":n" is a valid display name.  Instead, check for the
optional ".m" part and if it matches, just tack the complete $DISPLAY value
(including the leading colon) onto the hostname.

Fixes XOpenDisplay-10 when $DISPLAY is missing the optional screen number part.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>