[dist] Build mc for dist
mc is used to compile the .m files into .c files at dist time.  If
dist is run from a clean build directory, mc is not built and the dist
fails in xts5/Xproto when it tries to run it.

Add a dependency on mc for dist-hook to make sure that mc is built
before the tests.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>