[dist] Distribute files necessary to rebuild after distclean
The 'distclean' target deletes all of the .c files that were generated
during the dist step.  Since the .m files are not shipped in the
tarball, it becomes impossible to rebuild after that.

Fix this by shipping the .m files in the tarball along with the .mc
helper libraries that mc needs to rebuild the .c files.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>