xts5: Bail out with UNTESTED for BAD_LENGTH requests that would have length 0
The pBell-2 test sets the test type to BAD_LENGTH, which attempts to
create a request with a size that is too small by one long (4 bytes).
However, it tries to do this with an xBellReq, which is exactly 4
bytes in size.  This results in a request with a length field of 0,
which triggers the big request handling in the X server.  This makes
the server interpret the next request as the size part of the Bell
request, and hilarity ensues.

Instead, detect this case and bail out with an UNTESTED result.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>