Decouple the test suite from "make check"
The "check" target is run during "distcheck" after the test suite is built, so
having it run the test suite itself means that the suite has to pass before a
tarball can be distributed.  Also, an unsuspecting developer attempting to
create a tarball might accidentally run the test suite against his host X
server, which is bad because some of the tests can have adverse effects.

Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Acked-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Dan Nicholson <dbn.lists@gmail.com>