xts5: XWriteBitmapFile-3: Report UNTESTED instead of FAIL when run as root
This test creates a directory, makes it non-writable, then calls
XWriteBitmapFile to attempt to write to it, expecting it to fail with
BitmapOpenFailed.  This works fine when run as a normal user, but when
run as root, writing the file succeeds anyway.

Check the euid and bail out with UNTESTED if it's root.

Reported-by: Abhinav Shukla <abhinavs@nvidia.com>
Signed-off-by: Aaron Plattner <aplattner@nvidia.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>