Xlib4: Fix XMapSubwindows test 9
XMapSubwindows-9 creates a number of overlapping sibling windows, then
checks that a GetImage of each window is filled after processing all
Expose events.

This is invalid per the spec, as areas of a GetImage that are obscured
by a sibling are undefined. In particular, current servers intentionally
blank any area obscured by a sibling for security reasons.

Change XMapSubwindows to only verify the parts of the GetImage that are
not obscured by a sibling window.

Signed-off-by: Peter Harris <pharris@opentext.com>