Disable ColormapNotify test if implicit colormaps
If the server supports more than one simultaneously installed colormap,
it probably does implicit installation too. If the server implicitly
installs the colormap when it is created, there will be no
ColormapNotify when InstallColormap is called. In that case, the event
count will be wrong, and the test will report FAIL even though it
shouldn't.

Signed-off-by: Peter Harris <pharris@opentext.com>